requirements.txt
aiofiles==0.4.0
aioredis==1.2.0
appdirs==1.4.3
async-timeout==3.0.1
asyncpg==0.18.3
attrs==19.1.0
black==19.3b0
blinker==1.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.7
h11==0.8.1
h2==3.1.0
hiredis==1.0.0
hpack==3.0.0
httptools==0.0.13
Hypercorn==0.6.0
hyperframe==5.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.5.2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyNaCl==1.3.0
pytoml==0.1.20
Quart==0.9.0
requests==2.21.0
six==1.12.0
sortedcontainers==2.1.0
toml==0.10.0
typing-extensions==3.7.2
ulid-py==0.0.9
urllib3==1.24.3
uvicorn==0.7.1
uvloop==0.12.2
websockets==7.0
wsproto==0.14.0
